feat(api): snapshot cache for pressure sensitivity page
Moves pressure sensitivity computation behind an assumption_analysis_snapshots
cache, following the same pattern as domain analysis.

- snapshot-builder.ts: moves all computation logic (validation, transcript
  streaming, cell accumulation) out of the resolver; adds preparePressureSensitivityState
  (input hash), buildPressureSensitivitySnapshotOutput (all-models), writeSnapshot
- snapshot-cache.ts: FRESH/STALE/MISSING cache lifecycle; STALE queues a background
  rebuild and serves cached data; MISSING computes synchronously; filters
  modelIds/providerId at serve time rather than at compute time
- refresh-pressure-sensitivity-snapshot.ts: PgBoss worker that calls refreshPressureSensitivitySnapshot
- queue/types.ts: adds refresh_pressure_sensitivity_snapshot job type (10 min expiry)
- handler-config.ts: registers the new handler
- pressure-sensitivity.ts resolver: 965 lines → 25 lines
