+++ cloud/apps/api/src/services/pressure-sensitivity/snapshot-types.ts
@@ -1,0 +1,3 @@
+export const PRESSURE_SENSITIVITY_SNAPSHOT_TYPE = 'pressure_sensitivity';
+export const PRESSURE_SENSITIVITY_SNAPSHOT_CODE_VERSION = '1.0.0';
+export const PRESSURE_SENSITIVITY_ASSUMPTION_PREFIX = 'pressure-sensitivity';
