+++ cloud/apps/api/tests/queue/handlers/index.test.ts
@@ -51,8 +51,9 @@ describe('Handler Registration', () => {
       expect(jobTypes).toContain('analysis_result_janitor');
       expect(jobTypes).toContain('aggregate_analysis');
       expect(jobTypes).toContain('refresh_domain_analysis_snapshot');
+      expect(jobTypes).toContain('refresh_pressure_sensitivity_snapshot');
       expect(jobTypes).toContain('top_up_probes');
-      expect(jobTypes).toHaveLength(12);
+      expect(jobTypes).toHaveLength(13);
     });
   });
 
@@ -103,7 +104,7 @@ describe('Handler Registration', () => {
 
       // Should still work
       const jobTypes = getJobTypes();
-      expect(jobTypes).toHaveLength(12);
+      expect(jobTypes).toHaveLength(13);
     });
   });
